Show solutions in recyclerview.

+++ app/src/main/java/com/medicalrecord/utils/CustomViewModelFactory.kt
@@ -5,7 +5,7 @@ import android.arch.lifecycle.ViewModel
 import android.arch.lifecycle.ViewModelProvider
 import com.medicalrecord.data.viewmodels.CalculationViewModel
 
-class CustomViewModelFactory(private val application: Application, private val patientId: Int) : ViewModelProvider.NewInstanceFactory() {
+class CustomViewModelFactory(private val application: Application, private val patientId: Long) : ViewModelProvider.NewInstanceFactory() {
 
     override fun <T : ViewModel?> create(modelClass: Class<T>): T {
         return CalculationViewModel(application, patientId) as T
